% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newyorktimes.R
\name{best_recommend}
\alias{best_recommend}
\title{Find most recommend comments.}
\usage{
best_recommend(your_url)
}
\arguments{
\item{your_url}{A Character.}
}
\value{
the most recommend user comments based on \code{your_url}.
}
\description{
Get the most recommend comments based on article url
}
\examples{
best_recommend('https://www.nytimes.com/2019/12/08/opinion/donald-trump-impeachment.html')
}
