% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newyorktimes.R
\name{comments}
\alias{comments}
\title{Get Comments.}
\usage{
comments(your_url)
}
\arguments{
\item{your_url}{A Character.}
}
\value{
A dataframe include user id, user name, user location, and user comments based on \code{your_url}.
}
\description{
Get article's top 25 comments based on article url, and return a dataframe include user id, user name, user location, and user comments.
}
\examples{
comments('https://www.nytimes.com/2019/12/08/opinion/donald-trump-impeachment.html')
}
