% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newyorktimes.R
\name{mostpopular}
\alias{mostpopular}
\title{Search Most Popular Articles.}
\usage{
mostpopular(type = "emailed", period = 1)
}
\arguments{
\item{type}{A Character: choose from emailed, shared, viewed.}

\item{period}{A Numeric number: choose from 1,7,30.}
}
\value{
A dataframe inclde article's category, title, abstract, url, author, published date based on \code{topic} and \code{searchdate}.
}
\description{
Get the most popular articles in specifc type and recent 1 day, 7 day or 30 day and put the information of most popular article's url, category, author, title, abstract, published date in a dataframe.
}
\examples{
mostpopular('emailed', 1)

}
