% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyt_solve.R
\name{nyt_solve}
\alias{nyt_solve}
\title{Solve the Spelling Bee Challenge}
\usage{
nyt_solve(bee_letters, middle_letter)
}
\arguments{
\item{bee_letters}{the list of letters to use}

\item{middle_letter}{the middle letter which must always be included}
}
\value{
a list of solutions in descending order of word length
}
\description{
Solve the Spelling Bee Challenge
}
\examples{
nyt_solve(letters = "iloveyo", middle_letter = "o")
}
