% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newyorktimes.R
\name{search_article}
\alias{search_article}
\title{Search Article Based on Topic.}
\usage{
search_article(topic = "new york", searchdate = 20191231)
}
\arguments{
\item{topic}{A Character.}

\item{searchdate}{A numeric number in date formate: yearmonthdate.}
}
\value{
A dataframe inclde article's category, headline, abstract, lead paragraph, url, published year, published month, published day, total words of article based on \code{topic} and \code{searchdate}.
}
\description{
Get the top 10 articles' information based on topic and searchdate and return article's category, headline, abstract, lead paragraph, url, published year, published month, published day, total words of article in a dataframe.
}
\examples{
search_article('new york', 20191231)

}
