% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newyorktimes.R
\name{specific_articles}
\alias{specific_articles}
\title{Search ALL Article.}
\usage{
specific_articles(year = 2019, month = 1)
}
\arguments{
\item{year}{A series of Character.}

\item{month}{A numeric number in date formate: yearmonthdate.}
}
\value{
A dataframe inclde article's category, head, url, published date, total words of article based on \code{year} and \code{month}.
}
\description{
Get all articles in specific year and month and get the information: article's category, head, url, published date, total words of article in a dataframe.
}
\examples{
specific_articles(2019, 1)

}
